mach-exynos: stand-hotplug: secondcore app support + enabling/disabling hotplug logic based on the selected governor.
thanks to tegrak for secondcore app and the original modifications for the source code and thanks to pulser for porting it to s2.
in courtesy to the original author the comments are not removed. the original implementation is modified a little to use second_core_on variable as the number of cores in hotplugging (1: hotplug_off and single core mode, >2: the cores will be either always on or hotplug depending on the hotplug_on setting)

and also, this hotplugging logic will be disabled if a hotplug-aware governor is selected (pegasusq and hotplug governors) and enabled if any other governor is selected.

Signed-off-by: netchip <netchip@netchip-ubuntu.(none)>